+++ src/app/authloading/page.tsx
@@ -1,38 +1,24 @@
-import { createServerComponentClient } from '@supabase/auth-helpers-nextjs';
-import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
+'use client';
 
-import type { Database } from '@/types/supabase';
-// authloading페이지에서 session이 null 이라 아무것도 되지 않음
-export default async function AuthLoading() {
-  const supabaseServerClient = createServerComponentClient<Database>({ cookies });
+import React from 'react';
 
-  const {
-    data: { session },
-  } = await supabaseServerClient.auth.getSession();
+import { useRouter } from 'next/navigation';
 
-  if (session) {
-    const {
-      id,
-      email,
-      user_metadata: { name: nickname },
-    } = session.user;
+import { checkGoogleUser } from '@/api/auth';
+import Loading from '@/components/common/loading/Loading';
 
-    const { data: check } = await supabaseServerClient.from('users').select('id').eq('id', id);
-    if (check && check.length === 0) {
-      const user = {
-        id,
-        email: email as string,
-        nickname,
-      };
+export default function AuthLoading() {
+  const router = useRouter();
 
-      await supabaseServerClient.from('users').insert(user);
-    }
+  React.useEffect(() => {
+    const check = async () => {
+      await checkGoogleUser();
+      router.replace('/main');
+      router.refresh();
+    };
 
-    redirect('/main');
-  }
+    check();
+  }, []);
 
-  // setVisibilitySideBar(true);
-
-  return null;
+  return <Loading full={true} />;
 }
